Brought back sphinx indices to app/indices, moved ES to app/es_indices; turned ActsAsElasticModel into ActiveSupport Conern
