+++ spec/lib/elastic_model/elastic_model_spec.rb
@@ -1,4 +1,4 @@
-require File.expand_path("../../spec_helper", __FILE__)
+require "spec_helper"
 
 describe ElasticModel do
