+++ spec/lib/elastic_model/indices/taxon_index_spec.rb
@@ -1,0 +1,9 @@
+require "spec_helper"
+
+describe "Taxon Index" do
+  it "as_indexed_json should return a hash" do
+    t = Taxon.make!
+    json = t.as_indexed_json
+    expect( json ).to be_a Hash
+  end
+end
